submissions/Dia-Vats/level6/requirements.txt
streamlit>=1.35.0
neo4j>=5.19.0
pandas>=2.2.0
plotly>=5.22.0
python-dotenv>=1.0.1
numpy>=1.26.0